requirements.txt
streamlit>=1.28.0
yfinance>=0.2.28
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
joblib>=1.3.0
PyJWT>=2.8.0
bcrypt>=4.0.0
python-dotenv>=1.0.0

